+++ ccSchwabManagerTests/UtilitiesTests/SecretsTests/testKeychainManager.swift
@@ -9,6 +9,7 @@ import Testing
 import Foundation
 @testable import ccSchwabManager
 
+@MainActor
 struct testKeychainManager
 {